backend/requirements.txt
# ── Core ──
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
PyYAML==6.0.2

# ── Database ──
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
pgvector>=0.3.0
neo4j>=5.0.0
alembic>=1.13.0

# ── LLM & AI ──
anthropic>=0.42.0
google-genai>=1.0.0
openai>=1.50.0

# ── Data Processing ──
pandas==2.2.3
openpyxl==3.1.5
pdfplumber>=0.10.0
python-multipart>=0.0.9

# ── Export ──
python-pptx>=1.0.0

# ── Auth ──
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# ── Infra ──
redis>=5.0.0
sse-starlette>=2.0.0
httpx>=0.27.0
